requirements.txt
pinecone-client
pinecone-datasets
python-dotenv
sentence-transformers